refactor : refactor project directory and add comments
